LPS-144178 search-experiences-web: Keep sidebar organization consistent by moving AddSXPElementSidebar
