+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/EditSXPBlueprintForm.js
@@ -49,7 +49,7 @@ import {
 	validateNumberRange,
 	validateRequired,
 } from '../utils/validation';
-import AddSXPElementSidebar from './AddSXPElementSidebar';
+import AddSXPElementSidebar from './add_sxp_element_sidebar';
 import ClauseContributorsSidebar from './clause_contributors_sidebar/index';
 import ConfigurationTab from './configuration_tab/index';
 import PreviewSidebar from './preview_sidebar/index';
